+++ agents/kafka/src/agent/zk.rs
@@ -8,7 +8,6 @@ use std::time::Duration;
 use failure::ResultExt;
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
-use serde_json;
 use slog::debug;
 use slog::error;
 use slog::info;